+++ lib/ComposerInspector/Package.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Phpactor\ComposerInspector;
+
+class Package
+{
+    public function __construct(public string $name, public string $version, public bool $isDev)
+    {
+    }
+}