+++ lib/Configurator/Model/Change.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Phpactor\Configurator\Model;
+
+interface Change
+{
+    public function prompt(): string;
+}